requirements.txt
Django >= 2
pillow
psycopg2
psycopg2-binary
python-decouple
djangorestframework
pygments